Loader as component
