+++ client/src/Components/PrivateRoutes/LoginRoute/index.jsx
@@ -1,11 +1,9 @@
-import React from 'react';
-import { Route, Redirect } from 'react-router-dom';
-import {Authentication} from '../../../App';
+// import React from 'react';
+// import { Route, Redirect } from 'react-router-dom';
+// import AuthLoader from '../../Authenthication'
 
-const LoginRoute = ({ component: Component, ...rest }) => {
-  return (
-    <Route {...rest} render={props => (Authentication()  ?  <Redirect to="/"  /> : <Component {...props} />)} />
-  );
-};
+// const LoginRoute = ({ component: Component, ...rest }) => {
+//   return <Route {...rest} render={props => (<AuthLoader></AuthLoader> ? <Redirect to="/" /> : <Component {...props} />)} />;
+// };
 
-export default LoginRoute;
+// export default LoginRoute;
