+++ client/src/Components/PrivateRoutes/PrivateRoute/index.jsx
@@ -1,9 +1,9 @@
-import React from 'react';
-import { Route, Redirect } from 'react-router-dom';
-import { Authentication } from '../../../App';
+// import React from 'react';
+// import { Route, Redirect } from 'react-router-dom';
+// import AuthLoader from '../../Authenthication'
 
-const PrivateRoute = ({ component: Component, ...rest }) => {
-  return <Route {...rest} render={props => (Authentication()  ? <Component {...props} /> : <Redirect to="/login" />)} />;
-};
+// const PrivateRoute = ({ component: Component, ...rest }) => {
+//   return <Route {...rest} render={props => (<AuthLoader></AuthLoader> ? <Component {...props} /> : <Redirect to="/login" />)} />;
+// };
 
-export default PrivateRoute;
+// export default PrivateRoute;
