+++ client/src/utils/getToken.jsx
@@ -1,0 +1,5 @@
+ const getToken = () => {
+    return localStorage.getItem('status');
+}
+
+export default getToken
